decoupled moe from qwen omni specific modules
